pyproject.toml
[project]
name = "ai-coding-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai==1.12.1",
    "langchain-huggingface>=1.2.0",
    "openai>=2.14.0",
    "python-dotenv==1.1.0",
]
